Progress Made

+++ server/src/main/java/service/Service.java
@@ -6,6 +6,7 @@ import dataaccess.AuthDAO;
 import dataaccess.GameDAO;
 import dataaccess.UserDAO;
 import model.AuthData;
+import model.GameData;
 import model.UserData;
 import server.*;
 
@@ -25,7 +26,7 @@ public class Service {
 
     public RegisterResult register(UserData userInfo)
     {
-        Collection<String> userData = userDAO.getUser(userInfo.username());
+        UserData userData = userDAO.getUser(userInfo.username());
 
         // Verify the Provided Username Does Not Exist
         if(userData != null) {
@@ -46,7 +47,7 @@ public class Service {
 
     public LoginResult login(String username, String password) {
 
-        Collection<String> userInfo = userDAO.getUser(username);
+        UserData userInfo = userDAO.getUser(username);
 
         // Verify Username Exists
         if(userInfo == null) {
@@ -54,7 +55,7 @@ public class Service {
         }
 
         // Verify Password is Correct
-        if(!userInfo.contains(password)) {
+        if(Objects.equals(userInfo.password(), password)) {
             throw new ServiceException(401, "Error: Incorrect Password");
         }
 
@@ -69,6 +70,12 @@ public class Service {
 
     public LogoutResult logout(String authToken) {
 
+        // Verify Authentication
+        String validAuth = authDAO.getAuth(authToken);
+        if (!Objects.equals(authToken, validAuth)) {
+            throw new ServiceException(401, "Error: Unauthorized Access");
+        }
+
         // Delete Authentication Token
         authDAO.deleteAuth(authToken);
 
@@ -77,21 +84,46 @@ public class Service {
 
     public ListGamesResult listGames(String authToken) {
 
-        HashMap<String, Collection<String>> games = gameDAO.getGames();
+        // Verify Authentication
+        String validAuth = authDAO.getAuth(authToken);
+        if (!Objects.equals(authToken, validAuth)) {
+            throw new ServiceException(401, "Error: Unauthorized Access");
+        }
+
+        HashMap<String, GameData> games = gameDAO.listGames();
         return new ListGamesResult(games);
     }
 
     public CreateGameResult createGame(String gameName, String authToken) {
 
+        // Verify Authentication
+        String validAuth = authDAO.getAuth(authToken);
+        if (!Objects.equals(authToken, validAuth)) {
+            throw new ServiceException(401, "Error: Unauthorized Access");
+        }
+
         // Create New Game
-        String gameID = gameDAO.createGame(gameName);
+        Integer gameID = gameDAO.createGame(gameName);
 
         return new CreateGameResult(gameID, null, null, gameName);
     }
 
-//    public JoinGameResult joinGame() {
-//        return new JoinGameResult();
-//    }
+    public JoinGameResult joinGame(String playerColor, String authToken, Integer gameID) {
+
+        // Verify Authentication
+        String validAuth = authDAO.getAuth(authToken);
+        if (!Objects.equals(authToken, validAuth)) {
+            throw new ServiceException(401, "Error: Unauthorized Access");
+        }
+
+        String username = "?";
+
+        // Join Game
+        gameDAO.joinGame();
+
+
+        return new JoinGameResult();
+    }
 
     public ClearResult clear() {
